match.funfn {gsubfn}	R Documentation

Generic extended version of R match.fun

Description

A generic match.fun.

Usage

match.funfn(FUN, descend = TRUE)

Arguments

FUN
Function, character name of function or formula describing function.

descend
logical; control whether to search past non-function objects.

Details

The default method is the same as match.fun and the formula method is the same as as.function.formula. This function can be used within the body of a function to convert a function specification whether its a function, character string or formula into an actual function.

Value

Returns a function.

See Also

See Also match.fun, as.function.formula.

Examples

# return first argument evaluated at second argument squared.
sq <- function(f, x) {
   f <- match.funfn(f)
   f(x^2)
}

# call sq using different forms for function
sq(function(x) exp(x)/x, pi)
f <- function(x) exp(x)/x
sq("f", pi)
sq(~ exp(x)/x, pi)
sq(x ~ exp(x)/x, pi)

[Package gsubfn version 0.6-6 Index]