read.pattern {gsubfn}	R Documentation

Read file or text string using a regular expression to separate fields.

Description

Read file or text string using a regular expression to separate fields.

Usage

read.pattern(file, pattern, perl = FALSE, text, sep = "\01", fileEncoding = "", ...)

Arguments

file
See read.table.

pattern
A regular expression.

perl
logical. If TRUE then a perl style regular expression is expected.

text
See read.table.

sep
This should be a character that does not appear in any field. It is used for holding an intermediate form of the text.

fileEncoding
See read.table.

...
Further arguments passed to read.table.

Details

The parenthesized portions of the regular expression are extracted as fields.

Value

A data.frame.

See Also

See strapplyc, read.table.

Examples

# 3 space-separated fields.  The first field may have embedded spaces.
Lines <- "this is the first field 1 2
more text 3 4
"
pat <- "^(.*) +(\\S+) +(\\S+)$"
read.pattern(text = Lines, pattern = pat, as.is = TRUE)

# 3 space-separated fields.  The last field may have embedded spaces.
Lines2 <- "1 2 this is the first field
3 4 more text
"
pat2 <- "^(\\S+) +(\\S+) +(.*)$"
read.pattern(text = Lines2, pattern = pat2, as.is = TRUE)

[Package gsubfn version 0.6-6 Index]